# Use libosmium for geometry building