--- a/contrib/libosmium/README.contrib
+++ b/contrib/libosmium/README.contrib
@@ -1,2 +1,2 @@
 Source: https://github.com/osmcode/libosmium
-Revision: v2.11.0
+Revision: ab8236405330607932c8f6395956ec6d9bcdee5c