--- a/contrib/libosmium/osmium/area/detail/segment_list.hpp
+++ b/contrib/libosmium/osmium/area/detail/segment_list.hpp
@@ -110,16 +110,16 @@ namespace osmium {
                     });
                 }
 
-                uint32_t extract_segments_from_way_impl(
-                    osmium::area::ProblemReporter *problem_reporter,
-                    const osmium::Way &way, role_type role,
-                    bool ignore_invalid = false)
-                {
-                    uint32_t duplicate_nodes = 0;
+                uint32_t extract_segments_from_way_impl(osmium::area::ProblemReporter* problem_reporter, uint64_t& duplicate_nodes, const osmium::Way& way, role_type role) {
+                    uint32_t invalid_locations = 0;
 
                     osmium::NodeRef previous_nr;
                     for (const osmium::NodeRef& nr : way.nodes()) {
-                        if (ignore_invalid && !nr.location().valid()) {
+                        if (!nr.location().valid()) {
+                            ++invalid_locations;
+                            if (problem_reporter) {
+                                problem_reporter->report_invalid_location(way.id(), nr.ref());
+                            }
                             continue;
                         }
                         if (previous_nr.location()) {
@@ -135,7 +135,7 @@ namespace osmium {
                         previous_nr = nr;
                     }
 
-                    return duplicate_nodes;
+                    return invalid_locations;
                 }
 
             public:
@@ -220,29 +220,19 @@ namespace osmium {
                  * same node or different nodes with same location) are
                  * removed after reporting the duplicate node.
                  */
-                uint32_t extract_segments_from_way(
-                    osmium::area::ProblemReporter *problem_reporter,
-                    const osmium::Way &way, bool ignore_invalid = false)
-                {
+                uint32_t extract_segments_from_way(osmium::area::ProblemReporter* problem_reporter, uint64_t& duplicate_nodes, const osmium::Way& way) {
                     if (way.nodes().empty()) {
                         return 0;
                     }
                     m_segments.reserve(way.nodes().size() - 1);
-                    return extract_segments_from_way_impl(problem_reporter, way,
-                                                          role_type::outer,
-                                                          ignore_invalid);
+                    return extract_segments_from_way_impl(problem_reporter, duplicate_nodes, way, role_type::outer);
                 }
 
                 /**
                  * Extract all segments from all ways that make up this
                  * multipolygon relation and add them to the list.
                  */
-                uint32_t extract_segments_from_ways(
-                    osmium::area::ProblemReporter *problem_reporter,
-                    const osmium::Relation &relation,
-                    const std::vector<const osmium::Way *> &members,
-                    bool ignore_invalid = false)
-                {
+                uint32_t extract_segments_from_ways(osmium::area::ProblemReporter* problem_reporter, uint64_t& duplicate_nodes, const osmium::Relation& relation, const std::vector<const osmium::Way*>& members) {
                     assert(relation.members().size() >= members.size());
 
                     const size_t num_segments = get_num_segments(members);
@@ -251,18 +241,13 @@ namespace osmium {
                     }
                     m_segments.reserve(num_segments);
 
-                    uint32_t duplicate_nodes = 0;
-                    for_each_member(
-                        relation, members,
-                        [this, &problem_reporter, &duplicate_nodes,
-                         ignore_invalid](const osmium::RelationMember &member,
-                                         const osmium::Way &way) {
-                            duplicate_nodes += extract_segments_from_way_impl(
-                                problem_reporter, way,
-                                parse_role(member.role()), ignore_invalid);
-                        });
+                    uint32_t invalid_locations = 0;
+                    for_each_member(relation, members, [&](const osmium::RelationMember& member, const osmium::Way& way) {
+                        const auto role = parse_role(member.role());
+                        invalid_locations += extract_segments_from_way_impl(problem_reporter, duplicate_nodes, way, role);
+                    });
 
-                    return duplicate_nodes;
+                    return invalid_locations;
                 }
 
                 /**