--- a/contrib/libosmium/osmium/area/problem_reporter_exception.hpp
+++ b/contrib/libosmium/osmium/area/problem_reporter_exception.hpp
@@ -60,58 +60,64 @@ namespace osmium {
             ~ProblemReporterException() override = default;
 
             void report_duplicate_node(osmium::object_id_type node_id1, osmium::object_id_type node_id2, osmium::Location location) override {
-                m_sstream.str();
+                m_sstream.str("");
                 ProblemReporterStream::report_duplicate_node(node_id1, node_id2, location);
-                throw std::runtime_error(m_sstream.str());
+                throw std::runtime_error{m_sstream.str()};
             }
 
             void report_touching_ring(osmium::object_id_type node_id, osmium::Location location) override {
-                m_sstream.str();
+                m_sstream.str("");
                 ProblemReporterStream::report_touching_ring(node_id, location);
-                throw std::runtime_error(m_sstream.str());
+                throw std::runtime_error{m_sstream.str()};
             }
 
             void report_intersection(osmium::object_id_type way1_id, osmium::Location way1_seg_start, osmium::Location way1_seg_end,
                                      osmium::object_id_type way2_id, osmium::Location way2_seg_start, osmium::Location way2_seg_end, osmium::Location intersection) override {
-                m_sstream.str();
+                m_sstream.str("");
                 ProblemReporterStream::report_intersection(way1_id, way1_seg_start, way1_seg_end, way2_id, way2_seg_start, way2_seg_end, intersection);
-                throw std::runtime_error(m_sstream.str());
+                throw std::runtime_error{m_sstream.str()};
             }
 
             void report_duplicate_segment(const osmium::NodeRef& nr1, const osmium::NodeRef& nr2) override {
-                m_sstream.str();
+                m_sstream.str("");
                 ProblemReporterStream::report_duplicate_segment(nr1, nr2);
-                throw std::runtime_error(m_sstream.str());
+                throw std::runtime_error{m_sstream.str()};
             }
 
             void report_ring_not_closed(const osmium::NodeRef& nr, const osmium::Way* way = nullptr) override {
-                m_sstream.str();
+                m_sstream.str("");
                 ProblemReporterStream::report_ring_not_closed(nr, way);
-                throw std::runtime_error(m_sstream.str());
+                throw std::runtime_error{m_sstream.str()};
             }
 
             void report_role_should_be_outer(osmium::object_id_type way_id, osmium::Location seg_start, osmium::Location seg_end) override {
-                m_sstream.str();
+                m_sstream.str("");
                 ProblemReporterStream::report_role_should_be_outer(way_id, seg_start, seg_end);
-                throw std::runtime_error(m_sstream.str());
+                throw std::runtime_error{m_sstream.str()};
             }
 
             void report_role_should_be_inner(osmium::object_id_type way_id, osmium::Location seg_start, osmium::Location seg_end) override {
-                m_sstream.str();
+                m_sstream.str("");
                 ProblemReporterStream::report_role_should_be_inner(way_id, seg_start, seg_end);
-                throw std::runtime_error(m_sstream.str());
+                throw std::runtime_error{m_sstream.str()};
             }
 
             void report_way_in_multiple_rings(const osmium::Way& way) override {
-                m_sstream.str();
+                m_sstream.str("");
                 ProblemReporterStream::report_way_in_multiple_rings(way);
-                throw std::runtime_error(m_sstream.str());
+                throw std::runtime_error{m_sstream.str()};
             }
 
             void report_inner_with_same_tags(const osmium::Way& way) override {
-                m_sstream.str();
+                m_sstream.str("");
                 ProblemReporterStream::report_inner_with_same_tags(way);
-                throw std::runtime_error(m_sstream.str());
+                throw std::runtime_error{m_sstream.str()};
+            }
+
+            void report_invalid_location(osmium::object_id_type way_id, osmium::object_id_type node_id) override {
+                m_sstream.str("");
+                ProblemReporterStream::report_invalid_location(way_id, node_id);
+                throw std::runtime_error{m_sstream.str()};
             }
 
         }; // class ProblemReporterException