--- a/contrib/libosmium/osmium/io/detail/pbf_output_format.hpp
+++ b/contrib/libosmium/osmium/io/detail/pbf_output_format.hpp
@@ -33,7 +33,6 @@ DEALINGS IN THE SOFTWARE.
 
 */
 
-#include <algorithm>
 #include <cassert>
 #include <cmath>
 #include <cstdint>