This PR replaces the libgeos-based geometry builder with the libosmium area assembly. As libosmium approaches multipolygon assembly in a slightly different way, there are some visible changes to the polygon handling, as well as some secondary changes.

Changes in Polygon Handling
---------------------------

 * *Geometry type is determined by tagging only.* Previously osm2pgsql would create polygons only when a polygon was requested by the tagtransform and the geometry could actually assembled into rings. Otherwise a linestring was added to the *_ways table. This could lead to the odd situation where a partially broken multipolygon had some of its rings in the polygon table and the unclosed ways in the ways table. New behaviour is that only the result of tagtransform (or for output-multi: the requested format for the table) is taken into account. In some corner cases, this can lead to a regression with the C-transform when keys may describe a polygon or a way.
 * *Broken geometries are not recovered.* The libgeos geometry builder was very forgiving with broken polygons and would create a partial solution. The libosmium assembler uses an all or nothing approach. If the geometry is bad, the entire object is discarded. There is also no more buffering of geometries to recover them. On the plus side, libosmium can handle geometries where rings touch in more than one point, which is permissible in OSM but not for OGR polygons.
 * *`--exclude-invalid-polygon` removed.* libosmium never creates invalid geometries.
 * *Line merging has different results.* For route relations a custom line merging algorithm has been added which seems to be much more aggressive than the one employed by libgeos.

Other Changes
-------------

 * Reprojection of geometries now moved into geometry assembler because libosmium can build geometries only in WSG84.
 * Expiry works now on binary wkb only. This actually results in much cleaner code.
 * Some streamlining of middle interface (mostly working more with the native libosmium types).
 * now unused custom nodelist_t and multinodelist_t removed.

Todo
----

- [x] get libosmium area assembler changes into mainline (see https://github.com/osmcode/libosmium/issues/194)
- [x] fix regression tests (need to go each of the differences and check if the results are expected)
- [x] test with larger imports and visually check results
- [x] update documentation