@imagico 2017-02-06
Regarding broken geometries:

In the current state of libosmium w.r.t. assembling multipolygons, i.e. without https://github.com/osmcode/libosmium/issues/152 being addressed, this will have a huge impact on users in map rendering.  I have used libosmium quite extensively for data preparation for landcover rendering and there are areas where 20-50 percent of the big landcover polygons are broken in ways currently not recoverable by libosmium, mostly self intersections.  The visual impact on maps will be boldly visible.

That would be great for the standard style because it would bring down the number of broken geometries in the database quite quickly, faster than any QA tool can do, but it would certainly also be problematic for many osm2pgsql users.

Ideally libosmium should be able to handle these errors and you should be able to choose within osm2pgsql how strict you want to be about them.

Additional note:  If you assemble polygons in geographic coordinates and reproject them into other coordinate systems you can get invalid geometries.  This is rare practically with mercator but if you want to make sure you only have valid geometries you need a validity check afterwards.