@pnorman 2017-02-07
Appveyor is failing with 
```
C:\osm2pgsql\contrib\libosmium\osmium/tags/filter.hpp(41): fatal error C1083: Cannot open include file: 'boost/iterator/filter_iterator.hpp': No such file or directory
```

@alex85k, could this be something that's missing in the binary blobs used to supply the osm2pgsql dependencies?