@pnorman 2017-02-07
I've set up https://osmium.osm2pgsql.paulnorman.ca. The data on the right is from 170125 and imported with `./osm2pgsql -d osm2pgsql_test -S ../default.style --slim --drop --flat-nodes nodes.bin --cache 34000 ~/planet-170125.osm.pbf`