@lonvia 2017-02-07
The missing dependency for Appveyor is something we probably can (and would want to) get rid of in the code. So it's okay to leave it for now.